refactor(rgb): separate shared effects from device adapters
