+++ crates/lianli-media/src/rgb/sl_v4/twinkle.rs
@@ -1,5 +1,5 @@
 use super::engine::{brightness, palette_all, Color, Side};
-use crate::rgb::twinkle_pattern;
+use crate::rgb::effects::twinkle as twinkle_pattern;
 use lianli_shared::rgb::RgbEffect;
 
 pub(super) fn render(effect: &RgbEffect, fans: usize, _side: Side) -> Vec<Vec<Color>> {
