+++ crates/lianli-media/src/rgb/sync_effects/engine.rs
@@ -1,3 +1,4 @@
+pub(super) use crate::rgb::color::scale;
 use anyhow::{ensure, Result};
 use lianli_shared::rgb::{is_brightness_off, RgbEffect};
 
@@ -29,11 +30,8 @@ pub(super) fn palette(effect: &RgbEffect) -> [Color; 6] {
     colors
 }
 
-fn clamp_current(mut color: Color) -> Color {
-    while color.iter().map(|&channel| u16::from(channel)).sum::<u16>() > 600 {
-        color = color.map(|channel| (f64::from(channel) * 0.95) as u8);
-    }
-    color
+fn clamp_current(color: Color) -> Color {
+    crate::rgb::color::limit_current(color, 600)
 }
 
 pub(super) fn brightness(effect: &RgbEffect) -> u16 {
@@ -42,10 +40,6 @@ pub(super) fn brightness(effect: &RgbEffect) -> u16 {
     } else {
         [0, 64, 128, 192, 255][effect.brightness as usize]
     }
-}
-
-pub(super) fn scale(color: Color, factor: u16) -> Color {
-    color.map(|channel| ((u16::from(channel) * factor) >> 8) as u8)
 }
 
 pub(super) fn solid_frame(leds: usize, color: Color) -> Vec<Color> {
